for birch extractor; move the create_tf_train_feature and parse_tf_train_feature into two MixIn (depends when they generate list of passage or single passage per query at training time), so that they can be shared by each extractor as needed
